Move most of the logic from Consumer into a Cursor type.
